backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-generativeai>=0.5.0
pdfplumber>=0.11.0
numpy>=1.26.0
python-dotenv>=1.0.0
chromadb>=0.4.0